requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
appnope==0.1.2
argcomplete==1.10.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.11
boto3==1.35.13
botocore==1.35.17
cachetools==5.5.0
cairocffi==1.2.0
CairoSVG==2.5.2
catalogue==2.0.10
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==8.1.7
cloudpathlib==0.18.1
compressed_rtf==1.0.6
confection==0.1.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cssselect2==0.4.1
cymem==2.0.8
daphne==3.0.1
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.4
django-cors-headers==3.6.0
django-crontab==0.7.1
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.2
docx2txt==0.8
drf-yasg==1.20.0
ebcdic==1.1.1
et-xmlfile==1.1.0
extract-msg==0.28.7
filelock==3.15.4
fsspec==2024.6.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
gunicorn==19.6.0
hiredis==1.1.0
html5lib==1.1
huggingface-hub==0.24.2
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
IMAPClient==2.1.0
importlib_resources==6.4.0
incremental==17.5.0
inflection==0.5.1
ipdb==0.13.5
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.17.3
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==2.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mdurl==0.1.2
mmh3==3.0.0
mo-future==3.147.20327
moviepy==1.0.3
moz-sql-parser==3.32.20026
mpmath==1.3.0
msgpack==1.0.2
murmurhash==1.0.10
networkx==3.1
nltk==3.7
numpy==1.24.4
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.5
packaging==24.1
pandas==1.4.4
parso==0.8.1
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkgutil_resolve_name==1.3.10
preshed==3.0.9
proglog==0.1.9
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyfcm==1.5.1
Pygments==2.7.4
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.3.1
Pyphen==0.10.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.1
python-pptx==0.6.23
python_http_client==3.3.1
pytz==2020.5
PyYAML==6.0.1
redis==3.5.3
regex==2024.5.15
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.2.0
rsa==4.9
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sendgrid==6.4.8
sentence-transformers==3.0.1
service-identity==18.1.0
shellingham==1.5.4
six==1.12.0
smart-open==7.0.4
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
sqlparse==0.4.1
srsly==2.4.8
starkbank-ecdsa==1.1.0
sympy==1.13.1
textract==1.6.5
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.1.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.55.1
traitlets==5.0.5
transformers==4.43.2
twilio==6.56.0
Twisted==20.3.0
txaio==20.12.1
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==3.0.1
urllib3==1.26.2
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
WeasyPrint==52.5
webencodings==0.5.1
wget==3.2
wrapt==1.16.0
xlrd==1.2.0
XlsxWriter==3.0.1
zcrmsdk==3.0.0
zipp==3.19.2
zope.interface==5.2.0
